requirements.txt
ansi2html==1.8.0
appdirs==1.4.4
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
boto3==1.28.67
botocore==1.31.67
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
db-dtypes==1.1.1
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
Flask==2.2.5
fonttools==4.42.1
frozendict==2.3.8
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
greenlet==3.0.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
multitasking==0.0.11
nest-asyncio==1.5.7
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.17.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyodbc==5.0.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
retrying==1.3.4
rsa==4.9
s3transfer==0.7.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.22
SQLAlchemy-Utils==0.41.1
stack-data==0.6.2
tenacity==8.2.3
tornado==6.3.3
traitlets==5.10.0
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
wsproto==1.2.0
yfinance==0.2.28
zipp==3.17.0
